Add files via upload
Water Jug Problem.
You are given two jugs, a m-gallon one and a n-gallon one, a pump which has unlimited water which you can use to fill the jug, and the ground on which water may be poured. Neither jug has any measuring marking on it. How can you get exactly d-gallon of water in the m-gallon jug?